requirements.txt
streamlit>=1.38.0
snowflake-connector-python>=3.11.0
pandas>=2.2.2